Add meta heuristic and solution class

+++ read_data.py
@@ -11,7 +11,7 @@ def read(file):
     return data
 
 def find_segments(points):
-    distance=np.zeros((points.shape[0],points.shape[0]),dtype=int)
+    distance=np.zeros((points.shape[0],points.shape[0]))
     for i in range(points.shape[0]):
         for j in range (i+1, points.shape[0]):
             distance[i][j]=pow(pow(points[i][0]-points[j][0],2)+pow(points[i][1]-points[j][1],2),0.5)
